third_party/WebKit/LayoutTests/fast/viewport/viewport-40-expected.txt

ALERT: viewport size 700x352 scale 0.457143 with limits [0.457143, 5] and userScalable true
